chore: bump HackMyResume version to 2.0.0
With this upcoming release we've introduced potentially breaking functionality
so a major version bump is indicated. Additionally, it's been over a year since
the last HackMyResume release; a major version bump allows us to accept some
fruitful breakage, giving us a clean basis as we move forward on new features.